api/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
google-auth==1.19.0
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
gunicorn==20.0.4
psycopg2==2.8.5
requests==2.24.0
